Add dijkstra pathfinding algorithm
This adds one test and documentation regarding dijkstras pathfinding
algorithm. I'm not sure exactly how the documentation is generated,
but this adheres pretty strictly to the pattern set in the other
documentation so it should work.

This adds weighted undirected graphs to data structures as well, which
is a good standard implementation for testing the pathfinding algorithms.

* CONTRIBUTORS.md - Add myself
* docs/Data_Structure.rst - Add
+ docs/Pathing.rst - Add documentation for the new pathing/ submodule
* docs/index.rst - Add Pathing to the index
* pygorithm/__init__.py - Add myself to the contributors
* pygorithm/data_structures.py - Add WeightedUndirectedGraph and very minor
indentation changes
+ pygorithm/pathing/__init__.py - In the same format as the others
+ pygorithm/pathing/dijkstra.py - The simplest of pathfinding algorithms
* tests/test_data_structure.py - Add tests for the new graph type
+ tests/test_pathing.py - Test the example shown in the documentation