Add 1-dir A* pathfinding
This commit adds one-directional A* pathfinding; it also adds an
image to aid with one of the explanations. I wasn't 100% confident on
where these types of images should reside.

I also noticed this dijkstra pathfinding algorithm rechecks nodes in the
open list many times. I left it because dijkstra is mostly just an example
algorithm (though you do use similiar algorithms for precomputing paths).
I did change it to name the dict so it's slightly easier to read.

* docs/Pathfinding.rst - Add astar documentation
+ imgs/onedirectionalastar_riverexample.png - referenced in comments
* pygorithm/pathfinding/__init__.py - add astar
+ pygorithm/pathfinding/astar.py - add one-directional astar
* pygorithm/pathfinding/dijkstra - add comments on reverse_path
* tests/test_pathing - add timing tests. This is the same test as before
but such a way that it can be reused for all graph-based pathfinding
algorithms. This method won't work for smarter algorithms like Theta*
because they won't work on these style of graphs (there require additional
assumptions like a square or octogonal grid), but the TimedTestCase is
reusable.

I updated the test to check based on total weight rather than a specific
path which is more accurate to what is being tested. This indirectly
ensures that there is a way between each node in the path (otherwise
get_edge_weight would return None and an exception would be raised).