Attempt to use autoclass
This is an attempt to use autoclass for axis-aligned line. This should be
nearly parity and, to help compare, includes both the auto generated
documentation and the old style documentation.

This also fixes some non-command wording in axis-aligned line

* docs/Geometry.rst - Fix incorrect reST and minor wording improvements.
Also add autoclass for AxisAlignedLine.

* pygorithm/geometry/axisall.py - Add function skeletons and (hopefully)
something resembling parity to documentation in Geometry.rst