Axis-aligned line and line2 tests
Currently I'm running py 3.4.4 which doesn't have math.isclose. I'm not
sure the best place to put a similar method, right now I've replicated
it to the two locations that need it. Polygon will also need it. I'll try
to update python before pushing this upstream

* pygorithm/geometry/axisall.py - Implement everything, change intersects
return values and find_intersection return values (and update docs)

* pygorithm/geometry/line2.py - Implement everything. find_intersection
could be broken up but the subfunctions would need a lot of parameters
or it would repeat a lot of math

* tests/test_geometry.py - fix incorrect tests, use random.uniform instead
of random.randrange