Begin tests for extrapolated intersection
This uses a lot of images because generating tests for these is hard. The
image files are all very small (<5kb) so it shouldn't make downloading
the repo significantly slower.

This algorithm can be tricky to get right so I wanted to have a very
exhaustive test suite.

* pygorithm/geometry/extrapolated_intersection.py - add more helper
functions
* pygorithm/geometry/polygon2.py - fix fuzzing test sometimes failing
* tests/test_geometry.py - add regression test for polygon, add tests for
first two functions of extrapolated_intersection
+ imgs/test_geometry/test_extrapolated_intersection - images that help
visualize what the tests are doing. the naming convention is explained
in test_geometry.TestExtrapolatedIntersection docstring