Add tests and improve documentation on new sorting algorithms
This effectively standardizes the function arguments for all the new
sorting algorithms to be consistent with the other ones. Also fixes a few
bugs and performs some style improvements, such as moving documentation
of functions to after the function definition instead of before, and
adding documentation where it is missing.